Create TOSLibrary component
